+++ src/promise/js/extras/batch.js
@@ -1,17 +1,18 @@
-var slice = [].slice;
-
 /**
 Returns a new promise that will be resolved when all operations have completed.
-Takes both any numer of values as arguments. If an argument is a not a promise,
-it will be wrapped in a new promise, same as in `Y.when()`.
+Takes both callbacks and promises as arguments. If an argument is a callback,
+it will be wrapped in a new promise.
 
 @for YUI
 @method batch
-@param {Any} operation* Any number of Y.Promise objects or regular JS values
-@return {Promise} Promise to be fulfilled when all provided promises are
-                    resolved
+@param {Function|Promise} operation* Any number of functions or Y.Promise
+            objects
+@return {Promise}
+@deprecated @SINCE@
 **/
 Y.batch = function () {
+    Y.log('batch() was deprecated in YUI @SINCE@. Use Promise.every()', 'warn');
+
     var funcs     = slice.call(arguments),
         remaining = funcs.length,
         i         = 0,