+++ src/promise/js/promise.js
@@ -2,26 +2,8 @@
 Wraps the execution of asynchronous operations, providing a promise object that
 can be used to subscribe to the various ways the operation may terminate.
 
-When the operation completes successfully, call the Resolver's `fulfill()`
-method, passing any relevant response data for subscribers.  If the operation
-encounters an error or is unsuccessful in some way, call `reject()`, again
-passing any relevant data for subscribers.
-
-The Resolver object should be shared only with the code resposible for
-resolving or rejecting it. Public access for the Resolver is through its
-_promise_, which is returned from the Resolver's `promise` property. While both
-Resolver and promise allow subscriptions to the Resolver's state changes, the
-promise may be exposed to non-controlling code. It is the preferable interface
-for adding subscriptions.
-
-Subscribe to state changes in the Resolver with the promise's
-`then(callback, errback)` method.  `then()` wraps the passed callbacks in a
-new Resolver and returns the corresponding promise, allowing chaining of
-asynchronous or synchronous operations. E.g.
-`promise.then(someAsyncFunc).then(anotherAsyncFunc)`
-
-@module promise
-@since 3.9.0
+@module promise-core
+@since 3.11.0
 **/
 
 /**
@@ -67,7 +49,7 @@ function Promise(fn) {
     this._resolver = resolver;
 
     fn.call(this, function (value) {
-        resolver.fulfill(value);
+        resolver.resolve(value);
     }, function (reason) {
         resolver.reject(reason);
     });
@@ -88,22 +70,11 @@ Y.mix(Promise.prototype, {
                 resolves successfully
     @param {Function} [errback] function to execute if the promise
                 resolves unsuccessfully
-    @return {Promise} A promise wrapping the resolution of either "resolve" or
+    @return {Promise} A promise wrapping the resolution of either "fulfill" or
                 "reject" callback
     **/
     then: function (callback, errback) {
         return this._resolver.then(callback, errback);
-    },
-
-    /**
-    Returns the current status of the operation. Possible results are
-    "pending", "fulfilled", and "rejected".
-
-    @method getStatus
-    @return {String}
-    **/
-    getStatus: function () {
-        return this._resolver.getStatus();
     }
 });